WSTEAM1-262: Add higher order function to use MVT hook
